api/requirements.txt
Werkzeug==2.3.4
Flask==2.2.0
gunicorn==20.1.0
flask_cors==3.0.10
pandas==1.5.1
SQLAlchemy==1.4.44
psycopg2-binary==2.9.4
scikit-learn
flexpolyline
matplotlib
statsmodels